requirements.txt
streamlit>=1.32.0
google-genai>=1.0.0
psycopg2-binary>=2.9.9
pandas>=2.0.0
langfuse>=2.0.0
python-dotenv>=1.0.0
sqlparse>=0.5.0
